Refactor TasksAndModels into TaskInitializer and add doc
